requirements.txt
asgiref==3.2.10
attrs==19.3.0
autopep8==1.5.3
Django==3.0.8
django-hint==0.1.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-url-filters==0.5.1
flake8==3.8.3
importlib-metadata==1.7.0
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
voluptuous==0.11.7
wcwidth==0.2.5
zipp==3.1.0
